% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_index.R
\name{alpha_index}
\alias{alpha_index}
\title{Compute standard alpha indices}
\usage{
alpha_index(x, id = "sample")
}
\arguments{
\item{x}{a species matrix of observation/reads counts (samples as row,
OTU/ASV/species as columns)}

\item{id}{name to give to the output column containing the rownames}
}
\value{
A data.frame with the sample names and the three indices, each in a column.
}
\description{
Wrapper for species richness, Shannon and Simpson diversity indices formatted to a data-frame
}
