% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_residual.R
\name{alpha_residual}
\alias{alpha_residual}
\title{Linear model correction of sequencing depth bias for alpha diversity indices}
\usage{
alpha_residual(
  x,
  counts,
  id = "sample",
  index = "hill",
  trans = "auto",
  tree = NULL,
  scale = "scale",
  inform = F
)
}
\arguments{
\item{x}{is a species matrix of observation/reads counts (samples as row,
OTU/ASV/species as columns)}

\item{counts}{are untransformed per sample observation counts provided in a
named vector (same names as rownames of x). This parameter is optional,
if not provided, row sums of input matrix will be used.}

\item{id}{name to give to the output column containing the rownames of x}

\item{index}{either "hill" calling \code{\link{hill_index}} or "phylo" calling
\code{\link{phylo_index}}}

\item{trans}{define the desired read count's transformation. Either raw read
counts (trans="no"), square-root transformed (trans="sqrt") or log
transformed (trans="log"). Default is automatic selection for the best
fit transformation (trans="auto").}

\item{tree}{is needed for "phylo" index}

\item{scale}{the type of scaling to apply on residuals, either "no", "scale" for
center-scale with \code{\link[base:scale]{base::scale()}}, or "range" for
rescaling inside a fix range (-1:1) with \code{\link[scales:rescale]{scales::rescale()}}}

\item{inform}{if set to TRUE will print anova summary result of linear models
for each diversity index}
}
\value{
A data.frame in a long format with sample names, the transformed per
sample count of reads, the type of transformation used, the name of the
diversity index, the raw value of the index, and the scaled residual value
after linear model fitting against the transformed count of reads. If inform
is turn to TRUE, the summary of the anova of best linear model for each
index is printed in the terminal.
}
\description{
Compute residuals of linear model between alpha diversity indices (hill numbers
or phylogenetic alpha diversity) and sample counts from a species/OTU matrix
}
\seealso{
\code{\link[=hill_index]{hill_index()}} to compute Hill's numbers without correction
\code{\link[=phylo_index]{phylo_index()}} to compute phylogenetic alpha diversity indices without correction
}
