% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_pr2.R
\name{browse_pr2}
\alias{browse_pr2}
\title{Browse PR2 database taxonomy}
\usage{
browse_pr2(x, partial = TRUE)
}
\arguments{
\item{x}{a clade name}

\item{partial}{allow partial match, TRUE by default}
}
\value{
a rank name for exact match, a named list for partial match(es) or NA if not found
}
\description{
Determine the taxonomic rank of the provided clade name in the PR2 database
}
\references{
Guillou, L., Bachar, D., Audic, S., Bass, D., Berney, C., Bittner, L., Boutte, C. et al. 2013. The Protist Ribosomal Reference database (PR2): a catalog of unicellular eukaryote Small Sub-Unit rRNA sequences with curated taxonomy. Nucleic Acids Res. 41:D597–604.
}
