% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GuiRBioDiv.R
\docType{package}
\name{GuiRBioDiv}
\alias{GuiRBioDiv}
\title{Guillaume's collection of helper R function for biodiversity analyses}
\description{
Compute and normalize alpha diversity metrics, normalize variables and some more...
}
\author{
Guillaume Lentendu \email{guillaume.lentendu@unine.ch}
}
