% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interleave.R
\name{interleave}
\alias{interleave}
\title{Interleave}
\usage{
interleave(v1, v2, rev = F)
}
\arguments{
\item{v1}{a vector to interleave the first half with the second half,
or interleave with v2 if provided}

\item{v2}{a vector to interleave with v2}

\item{rev}{reverse the order of v2}
}
\value{
a vector with interleaved elements
}
\description{
Interleave elements of one or two vectors
}
