% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_exa.R
\docType{data}
\name{mat_exa}
\alias{mat_exa}
\title{Example ASV matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 10 rows and 1533 columns.
}
\usage{
data(mat_exa)
}
\description{
Example ASV matrix generated with V4 18S general primers (Lentendu \emph{et al}. 2023)
}
\references{
Lentendu, G., Lara, E., Geisen, S., 2023. Metabarcoding Approaches For Soil Eukaryotes, Protists and Microfauna, in: Martin, F., Uroz, S. (Eds.), Microbial Environmental Genomics (MEG), Methods in Molecular Biology. Springer, New York, NY, in press
}
\seealso{
\code{\link[=taxo_exa]{taxo_exa()}} for taxonomy of the ASVs

\url{https://github.com/lentendu/V4_SSU_ASV_bioinformatic_pipeline}
}
\keyword{datasets}
