% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permtab.R
\name{permtab}
\alias{permtab}
\title{Format PERMANOVA results}
\usage{
permtab(x, digits = 2)
}
\arguments{
\item{x}{an object of class anova.cca as produced by \code{\link[vegan:adonis2]{vegan::adonis2()}}}

\item{digits}{number of digits to round R^2 values}
}
\value{
A data.frame with the parameters the rounded R^2 and
stars denoting p-value significance (* for p<0.05, ** for
p<0.01, *** for p<0.001)
}
\description{
Transform PERMANOVA aov table as produced by \code{\link[vegan:adonis2]{vegan::adonis2()}}
into a tidy data-frame
}
\examples{
x<-data.frame(group=c("B","B","A","A","A","B","B","A","B","A"))
tmp<-vegan::adonis2(mat_exa ~ group, data = x)
permtab(tmp)

}
