% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_index.R
\name{phylo_index}
\alias{phylo_index}
\title{Phylogenetic alpha diversity indices}
\usage{
phylo_index(x, id = "sample", tre)
}
\arguments{
\item{x}{a species matrix of observation/reads counts (samples as row,
OTU/ASV/species as columns)}

\item{id}{name to give to the output column containing the rownames}

\item{tre}{phylogenetic tree (i.e. an object of class "phylo")}
}
\value{
A data.frame with the sample names and the three phylogenetic
indices,  each in a column.
}
\description{
Wrapper for three diversity metrics PD,  PSV and PSE for a species
matrix (species as columns) and a tree
}
