% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_exa.R
\docType{data}
\name{taxo_exa}
\alias{taxo_exa}
\title{Example taxonomy data-frame}
\format{
An object of class \code{data.frame} with 1533 rows and 19 columns.
}
\usage{
data(taxo_exa)
}
\description{
Example taxonomy data-frame related to ASVs generated with V4 18S general primers (Lentendu \emph{et al}. 2023)
}
\references{
Lentendu, G., Lara, E., Geisen, S., 2023. Metabarcoding Approaches For Soil Eukaryotes, Protists and Microfauna, in: Martin, F., Uroz, S. (Eds.), Microbial Environmental Genomics (MEG), Methods in Molecular Biology. Springer, New York, NY, in press
}
\seealso{
\code{\link[=mat_exa]{mat_exa()}} for the corresponding ASV read count matrix

\url{https://github.com/lentendu/V4_SSU_ASV_bioinformatic_pipeline}
}
\keyword{datasets}
