% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformTukey2.R
\name{transformTukey2}
\alias{transformTukey2}
\title{Tukey's Ladder of Powers transformation}
\usage{
transformTukey2(x, lambda = seq(-2, 2, 0.05), onlynn = FALSE)
}
\arguments{
\item{x}{a numeric vector to transform}

\item{lambda}{a numeric vector of exponents to screen in order
to find the transformation wit best normality}

\item{onlynn}{if TRUE, only perform the transformation if the
input vector is not already normal, otherwise return x}
}
\value{
a list with four elements:
\itemize{
\item val - a vector of transformed values
\item trans - the transformation applied
\item W - the value of the Shapiro-Wilk statistic of the transformed values
\item p - the p-value of the Shapiro-Wilk test of normality of the transformed values
}
}
\description{
Transform a numeric vector with Tukey's Ladder of Powers (also know
as Box-Cox transformation).
Code adapted from \code{\link[rcompanion:transformTukey]{rcompanion::transformTukey()}}
}
\examples{
set.seed(2)
transformTukey2(rnorm(10))

# avoid transformation if already normal
set.seed(2)
transformTukey2(rnorm(10), onlynn = TRUE)

}
\references{
http://rcompanion.org/handbook/I_12.html
}
